refactored collect_gt_utils.py
